Implement event triggering and additional callbacks
Plugin works properly when jquery.hotkeys is present; still requires
testing for use without it.

Triggers the custom event when the specified key is held for longer
than "duration"; if the key is released before that, the "onShortPress"
callback is executed. The "before" callback is executed before the
timeout is started.